games-util/steam-games-meta: Rework use-dependencies on libpng-1.2
This moves the dependency on libpng-1.2 from the portal USE flag to the
source USE flag because it is needed for at least Half-Life 2 as well
and thus probably for source games in general. To provide backwards
compatibility, the portal USE flag now depends on the source USE flag.

Furthermore, an abi_x86_32 build with USE dependency is added to
libpng-1.2 for AMD64. This is needed for 32-bit source games to start on
AMD64.