net-libs/libcurl-debian: New symlink package for Debian compatibility
This isn't going in the main tree, at least for now, as I'm a little
unsure about whether symlinking to different backends would ever
break. It doesn't appear to in most cases.

This doesn't deal with the fact that Debian also packages libcurl.so.3
and friends. Hopefully we won't need that.

Package-Manager: portage-2.3.0