games-util/esteam: Depending on subslots is badbadbad
Portage allows it but I have been pointed to the PMS, which says this
is not valid in ebuilds, and presumably Portage sets.

It's hard to know when to depend on SLOTs and when to depend on
versions (or both!) and a consistent rule cannot be applied because
packages vary a lot. I've tried my best here.

Package-Manager: portage-2.3.0