games-util/esteam: WormsWMD is unbundleable with lots of new entries
This game is especially annoying because it bundles FFmpeg 2.8 with
just about every optional feature enabled, meaning that the likes of
libbluray are also bundled. We do not have compatibility SLOTs for
FFmpeg although I will look into it. In the general case, it's also
hard to tell the difference between libraries that are really used by
a game and libraries that are there purely to satisfy unused features
so users will simply have to install these extra packages. Fortunately
I watch Blu-rays anyway. ;)

Package-Manager: Portage-2.3.64, Repoman-2.3.12
Signed-off-by: James Le Cuirot <chewi@gentoo.org>