games-util/esteam: Symlink libcurl.so.3 to libcurl.so.4
Yes, this is very controversial but, as I understand it, the API
changed while the ABI did not. There was some complication here around
OpenSSL 1.1 but it works for Dustforce regardless and until someone
reports that it breaks, it seems better than packaging a super
vulnerable curl 7.15.5.

Closes: https://github.com/anyc/steam-overlay/issues/253
Package-Manager: Portage-2.3.99, Repoman-2.3.22
Signed-off-by: James Le Cuirot <chewi@gentoo.org>