games-util/esteam: Write per-game lists into the set file
It was previously impossible to tell which games were pulling in which
dependencies. This writes each game name out as a comment, followed by
its dependencies. The duplicate entries do not matter.

It now looks a little untidy due to multilib duplicates within the
same game but this doesn't do any harm and addressing it doesn't seem
worth the effort.

Package-Manager: Portage-3.0.9, Repoman-3.0.1
Signed-off-by: James Le Cuirot <chewi@gentoo.org>