games-util/steam-launcher: Fix desktop layout for controllers in wayland
Upstream issue https://github.com/ValveSoftware/steam-for-linux/issues/4738

From what I understand Valve points to wayland, and wayland points to valve,
in the meantime this has been broken for way too long with no fix in sight.

The problem originates in the lack of the X11 XTEST extension in wayland
sessions.

Solution is:
- extest: this always works
- xwayland[libei] with a compositor that also supports libei (currently only
>=x11-wm/mutter-45.0 )

Since extest always works this is the first choice. If the library is present
the wrapper LD_PRELOAD's it as per upstream instructions.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>